Use clap 3.0
