+++ Cargo.toml
@@ -5,6 +5,7 @@ homepage = "https://toip.dev"
 license = "MIT"
 name = "toip"
 version = "0.1.0"
+description = "Tool to install other tools as containers but run them as if they are installed regularly."
 
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
@@ -12,7 +13,7 @@ version = "0.1.0"
 anyhow = "1.0"
 async-trait = "0.1.51"
 bytes = "1.0.1"
-clap-verbosity-flag = "0.3.1"
+clap-verbosity-flag = "0.4.0"
 directories = "4.0.1"
 flate2 = "1.0.20"
 futures-util = "0.3.15"
@@ -28,7 +29,7 @@ serde_json = "1.0.66"
 serde_yaml = "0.8.23"
 sha2 = "0.9.5"
 simplelog = "0.10.0"
-structopt = "0.3"
+clap = {version = "3.0.7", features = ["derive"]}
 tar = "0.4.37"
 termion = "1.5.6"
 thiserror = "1.0.26"
